requirements.txt
python-telegram-bot==22.1
openai==1.82.0